+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 10/21/2020 - boyd-mtu
 
 Organized project file paths, attempted to get calendar page working. Created test database interaction site. Edited readme to include instructions for docker. Created example db.ini file.
+
+11/14/2020 - ejbelt-mtu
+
++ Added functionality to calendar.js to make it display only the calendar itself, as well as making the edit and delete buttons functional
+
++ Added some functionality to findTimeGroup.php and findTimeSingle.php to allow for obtaining each user's time blocks.
+
++ Added findTime.js